# [chore] remove ie11 `on` fallback